+++ DBDao.go
@@ -102,7 +102,8 @@ type DataSourceConfig struct {
 
 	// FuncGlobalTransaction seata/hptx全局分布式事务的适配函数,返回IGlobalTransaction接口的实现
 	// 业务必须调用zorm.BindContextEnableGlobalTransaction(ctx)开启全局分布事务
-	FuncGlobalTransaction func(ctx context.Context) (IGlobalTransaction, context.Context, error)
+	// seata-go 的ctx是统一的绑定的是struct,也不是XID字符串.  hptx是分离的,所以返回了两个ctx,兼容两个库
+	FuncGlobalTransaction func(ctx context.Context) (IGlobalTransaction, context.Context, context.Context, error)
 	// DisableAutoGlobalTransaction 禁用自动全局分布式事务,默认false,虽然设置了FuncGlobalTransaction,但是并不想全部业务自动开启全局事务
 	// DisableAutoGlobalTransaction = false; ctx,_=zorm.BindContextEnableGlobalTransaction(ctx,false) 默认使用全局事务,ctx绑定为false才不开启
 	// DisableAutoGlobalTransaction = true;  ctx,_=zorm.BindContextEnableGlobalTransaction(ctx,true) 默认禁用全局事务,ctx绑定为true才开启
@@ -298,7 +299,7 @@ var transaction = func(ctx context.Context, doTransaction func(ctx context.Conte
 				globalTxOpen = true
 			}
 			// 获取分布式事务实现对象,用于控制事务提交和回滚.分支事务需要ctx中TX_XID有值,将分支事务关联到主事务
-			globalTransaction, globalRootContext, errGlobal = funcGlobalTx(ctx)
+			globalTransaction, ctx, globalRootContext, errGlobal = funcGlobalTx(ctx)
 			if errGlobal != nil {
 				errGlobal = fmt.Errorf("->Transaction-->global:Transaction FuncGlobalTransaction获取IGlobalTransaction接口实现失败:%w ", errGlobal)
 				FuncLogError(ctx, errGlobal)